+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\index.js":"1","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\reportWebVitals.js":"2"},{"size":500,"mtime":499162500000,"results":"3","hashOfConfig":"4"},{"size":362,"mtime":499162500000,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1w8jxkb",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\index.js",[],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\reportWebVitals.js",[]]
+[{"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\index.js":"1","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\reportWebVitals.js":"2","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\App.js":"3","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Root.js":"4","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Pages\\NextPage.js":"5","C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Pages\\HomePage.js":"6"},{"size":460,"mtime":1606386219769,"results":"7","hashOfConfig":"8"},{"size":362,"mtime":499162500000,"results":"9","hashOfConfig":"8"},{"size":519,"mtime":1606390165539,"results":"10","hashOfConfig":"8"},{"size":214,"mtime":1606386089845,"results":"11","hashOfConfig":"8"},{"size":2069,"mtime":1606387566517,"results":"12","hashOfConfig":"8"},{"size":2049,"mtime":1606390429429,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1w8jxkb",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":3,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":1,"source":null},{"filePath":"24","messages":"25","errorCount":2,"warningCount":5,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\index.js",[],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\reportWebVitals.js",[],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\App.js",[],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Root.js",[],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Pages\\NextPage.js",["26","27","28","29","30","31","32"],"C:\\Users\\zee20\\OneDrive\\문서\\GitHub\\chatapppractice\\src\\Pages\\HomePage.js",["33","34","35","36","37","38","39"],{"ruleId":"40","severity":1,"message":"41","line":1,"column":1,"nodeType":"42","messageId":"43","fix":"44"},{"ruleId":"45","severity":1,"message":"46","line":9,"column":10,"nodeType":"47","messageId":"48","endLine":9,"endColumn":22},{"ruleId":"45","severity":1,"message":"49","line":22,"column":10,"nodeType":"47","messageId":"48","endLine":22,"endColumn":18},{"ruleId":"50","severity":2,"message":"51","line":24,"column":25,"nodeType":"47","messageId":"52","endLine":24,"endColumn":34},{"ruleId":"50","severity":2,"message":"53","line":35,"column":35,"nodeType":"47","messageId":"52","endLine":35,"endColumn":39},{"ruleId":"50","severity":2,"message":"53","line":71,"column":34,"nodeType":"47","messageId":"52","endLine":71,"endColumn":38},{"ruleId":"54","severity":1,"message":"55","line":75,"column":13,"nodeType":"56","endLine":75,"endColumn":94},{"ruleId":"45","severity":1,"message":"57","line":1,"column":17,"nodeType":"47","messageId":"48","endLine":1,"endColumn":23},{"ruleId":"45","severity":1,"message":"58","line":1,"column":25,"nodeType":"47","messageId":"48","endLine":1,"endColumn":33},{"ruleId":"45","severity":1,"message":"59","line":10,"column":10,"nodeType":"47","messageId":"48","endLine":10,"endColumn":27},{"ruleId":"45","severity":1,"message":"60","line":26,"column":7,"nodeType":"47","messageId":"48","endLine":26,"endColumn":16},{"ruleId":"45","severity":1,"message":"61","line":27,"column":7,"nodeType":"47","messageId":"48","endLine":27,"endColumn":16},{"ruleId":"62","severity":2,"message":"63","line":79,"column":16,"nodeType":"64","endLine":79,"endColumn":20},{"ruleId":"62","severity":2,"message":"65","line":81,"column":12,"nodeType":"64","endLine":81,"endColumn":17},"unicode-bom","Unexpected Unicode BOM (Byte Order Mark).","Program","unexpected",{"range":"66","text":"67"},"no-unused-vars","'useAuthState' is defined but never used.","Identifier","unusedVar","'ChatRoom' is defined but never used.","no-undef","'firestore' is not defined.","undef","'auth' is not defined.","jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","'useRef' is defined but never used.","'useState' is defined but never used.","'useCollectionData' is defined but never used.","'firestore' is assigned a value but never used.","'analytics' is assigned a value but never used.","react/jsx-no-undef","'Link' is not defined.","JSXIdentifier","'Route' is not defined.",[-1,0],""]
